Provide better structure
